small code improvement

+++ Emulator/ppu.h
@@ -71,7 +71,7 @@ private:
 	uint8_t attribute_latch;
 	uint8_t prevoius_read_data;
 	uint8_t sprite_latch[8];// holds OAM attribute for sprite
-	uint8_t counter[8]; // holds x for sprite
+	int counter[8]; // holds x for sprite
 	// internal data
 	uint8_t dma_page;
 	uint8_t OAMaddr;
